requirements.txt
#Data Squad & Core Libraries
pandas
xarray
netcdf4
sqlalchemy
psycopg2-binary

#AI Squad Libraries
langchain
langchain-community
langchain-google-genai
sentence-transformers

#Frontend Squad Libraries
streamlit
plotly

fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
psycopg2-binary
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
netCDF4
xarray
pydantic